Add notes for lecture 1
I have chosen to add notes and stuff to a separate branch so that new students
can still simply fork the master branch and start the course.

Reference branch is supposed to be used only for well, reference. I have noted
that students struggle when they are just getting started to create notes (and
solutions), so a set of reference notes to match their own notes is helpful.

Intended use is that students first create their own notes, then the mentor
check their notes and instruct them to check the reference notes to fill the
gaps in their own notes.